[kuaikanmanhua] fixes
* fixes clipboard manga title
* accept mobile url
* fixes test filename (and test itself as chosen chapter was locked)
